Refactor main.py and add tests
Added .gitignore file.
Updated README file with instructions to run the tests.
Updated get_choices function in main.py to raise an error if invalid input
is entered.
Created a main function and added the conditional to only invoke it when
the main.py script is run directly.
Created test suite for the game.